Pmart/mart/requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
mysqlclient==2.2.4
pillow==10.4.0
PyMySQL==1.1.1
sqlparse==0.5.1
typing_extensions==4.12.2

pyotp~=2.9.0
djangorestframework-simplejwt~=5.3.1